feat: cleanup and improve topicId query
